# backEnd/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social Media API",
        "description": "Documentation for Social Media API",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/joueurs": {
            "get": {
                "tags": [
                    "Joueur"
                ],
                "summary": "Get all joueurs",
                "operationId": "eff393293a36f72473d724ddf12a50b6",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/JoueurResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Joueur"
                ],
                "summary": "Create a new joueur",
                "operationId": "3bf374d6eb0e1976cab122952e481c77",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Joueur created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JoueurResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/joueurs/{id}": {
            "get": {
                "tags": [
                    "Joueur"
                ],
                "summary": "Get a specific joueur",
                "operationId": "79c60566cf1667b3167058f3bd501515",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the joueur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JoueurResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Joueur not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/joueurs/{id}/experiences": {
            "get": {
                "tags": [
                    "Joueur"
                ],
                "summary": "Get experiences for a specific joueur",
                "operationId": "3ee0773e4e0e4938951ce41bf337fca5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the joueur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExperienceResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/experiences": {
            "post": {
                "tags": [
                    "Joueur"
                ],
                "summary": "Create a new experience for a specific joueur",
                "operationId": "73f3323fe422433ebd7b3f1b6a6f0d5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the joueur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nomClub",
                                    "endDate"
                                ],
                                "properties": {
                                    "nomClub": {
                                        "type": "string",
                                        "example": "Club de football"
                                    },
                                    "joinDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-01-01"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-01-01"
                                    },
                                    "place": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "categoryType": {
                                        "type": "string",
                                        "example": "SENIOR",
                                        "enum": [
                                            "SENIOR",
                                            "ESPOIR",
                                            "JUNIOR",
                                            "CADET",
                                            "MINIM"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Experience created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExperienceResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/joueurs/{id}/experiences/{experienceId}": {
            "put": {
                "tags": [
                    "Joueur"
                ],
                "summary": "Update a specific experience for a joueur",
                "operationId": "6b92f38e95ac0804f17f57ead660f21b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the joueur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "experienceId",
                        "in": "path",
                        "description": "ID of the experience",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nomClub": {
                                        "type": "string",
                                        "example": "Club de football"
                                    },
                                    "joinDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-01-01"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-01-01"
                                    },
                                    "place": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "categoryType": {
                                        "type": "string",
                                        "example": "SENIOR",
                                        "enum": [
                                            "SENIOR",
                                            "ESPOIR",
                                            "JUNIOR",
                                            "CADET",
                                            "MINIM"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Experience updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExperienceResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Experience not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Joueur"
                ],
                "summary": "Delete a specific experience for a joueur",
                "operationId": "ba77944b3ba4f54421abd2b9e4ccfd47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the joueur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "experienceId",
                        "in": "path",
                        "description": "ID of the experience",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Experience deleted successfully"
                    },
                    "404": {
                        "description": "Experience not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ExperienceResource": {
                "title": "ExperienceResource",
                "description": "Resource representing a player's experience",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nomClub": {
                        "type": "string",
                        "example": "FC Barcelona"
                    },
                    "joinDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2020-01-01"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-12-31"
                    },
                    "place": {
                        "type": "string",
                        "example": "Spain"
                    },
                    "categoryType": {
                        "type": "string",
                        "example": "Professional"
                    },
                    "joueur_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "JoueurResource": {
                "title": "JoueurResource",
                "description": "Resource representing a player",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "experiences": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ExperienceResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "User Resource",
                "description": "Resource representation of a User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "bio": {
                        "type": "string",
                        "example": "This is John's bio"
                    },
                    "profileImage": {
                        "type": "string",
                        "example": "http://localhost:8081/storage/profile_images/john.jpg"
                    },
                    "bannerImage": {
                        "type": "string",
                        "example": "http://localhost:8081/storage/banner_images/john.jpg"
                    },
                    "role": {
                        "type": "string",
                        "example": "Joueur"
                    },
                    "isActive": {
                        "type": "boolean",
                        "example": true
                    },
                    "stats": {
                        "properties": {
                            "followers_count": {
                                "type": "integer",
                                "example": 100
                            },
                            "following_count": {
                                "type": "integer",
                                "example": 50
                            },
                            "posts_count": {
                                "type": "integer",
                                "example": 20
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-02T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}